Consolidate performance scaling modules
The scaling feature is spread across seven Haskell modules. Collect its
implementation in PerfScaling and its tests in PerfScalingTests. Share
the command parser and pixel decoder used by the tests.
